Move the Loyverse integration into a dedicated integrations.loyverse folder, splitting the code into multiple files
