+++ integrations/loyverse/exceptions.py
@@ -1,0 +1,2 @@
+class InsufficientFundsError(Exception):
+    pass
