+++ modules/points.py
@@ -7,7 +7,9 @@ from telegram.ext import Application, CommandHandler, ContextTypes
 from helpers.access_checker import AccessChecker
 from helpers.exceptions import UserFriendlyError
 from helpers.points import Points
-from helpers.loyverse import LoyverseConnector, InsufficientFundsError
+
+from integrations.loyverse.api import LoyverseApi
+from integrations.loyverse.exceptions import InsufficientFundsError
 
 # sarcasm
 with open("resources/points_donate_sarcasm.txt", "r") as file:
@@ -18,8 +20,8 @@ with open("resources/points_balance_sarcasm.txt", "r") as file:
 
 
 class PointsModule:
-    def __init__(self, lc: LoyverseConnector, ac: AccessChecker):
-        self.lc = lc
+    def __init__(self, loy: LoyverseApi, ac: AccessChecker):
+        self.loy = loy
         self.ac = ac
 
     def install(self, application: Application) -> None:
@@ -32,7 +34,7 @@ class PointsModule:
             if not user:
                 raise UserFriendlyError("I don't really know who you are - to check your balance you first need to create a username in Telegram.")
 
-            balance = self.lc.get_balance(user)
+            balance = self.loy.get_balance(user)
             sarc = random.choice(balance_sarcastic_comments)
             await update.message.reply_text(f"{sarc} @{user}, you have {balance} T5 Loyalty Points!")
         except UserFriendlyError as e:
@@ -62,9 +64,9 @@ class PointsModule:
 
             try:
                 if not self.ac.can_donate_for_free(sender):
-                    self.lc.remove_points(sender, points)
+                    self.loy.remove_points(sender, points)
 
-                self.lc.add_points(recipient, points)
+                self.loy.add_points(recipient, points)
             except InsufficientFundsError as error:
                 raise UserFriendlyError("Your generosity is the stuff of legends, but you cannot donate more points than you have in your balance.") from error
             except Exception as error:
